Add modeling drone as an airborne vehicle
r5816
2021-06-13

Note: this commit does NOT make delivery drones fully
functional yet.

## DISCUSSION
1. https://github.com/mars-sim/mars-sim/issues/361

## NEW
1. Add Flyer and Drone class
2. Add Delivery Drone in vehicles.xml, settlements.xml,
   parts.xml, manufacturing.xml, and resupplies.xml.
3. Add Drone icon and svg images.
4. Add DroneDisplayInfoBean

## CHANGE
1. Designate names for drones as "Drone 001", "Drone 002", etc.
2. Use double for probability in MalfunctionManager,
   MalfunctionFactory, PartTableModel, MaintenanceScope,
   PartConfig.
3. Revise UnitDisplayInfoFactory to return drone bean.

## FIX
1. Add TabPanelEVA only when the vehicle is a rover in
   VehicleWindow.
   - Drone and LUV won't need TabPanelEVA.
2. Correct not being able to display LUV icon.
3. Remove NavigationTabPanel for LUV in VehicleWindow.
4. Add @Override tag to getButtonIcon() and getSound() in
   LUVDisplayInfoBean and DroneDisplayInfoBean.

+++ mars-sim-core/src/main/java/org/mars_sim/msp/core/resource/MaintenanceScope.java
@@ -15,7 +15,7 @@ public class MaintenanceScope implements Serializable {
 	// Domain members
 	private String scope;
 	private Part part;
-	private int probability;
+	private double probability;
 	private int maxNumber;
 
 	/**
@@ -26,7 +26,7 @@ public class MaintenanceScope implements Serializable {
 	 * @param probability the probability of this part being needed for maintenance.
 	 * @param maxNumber   the maximum number of this part needed for maintenance.
 	 */
-	public MaintenanceScope(Part part, String scope, int probability, int maxNumber) {
+	public MaintenanceScope(Part part, String scope, double probability, int maxNumber) {
 		if (scope == null) {
 			throw new IllegalArgumentException(Msg.getString("Part.error.nameIsNull")); //$NON-NLS-1$
 		}
@@ -40,7 +40,7 @@ public class MaintenanceScope implements Serializable {
 		return part;
 	}
 	
-	public int getProbability() {
+	public double getProbability() {
 		return probability;
 	}
